refactor: add useDarkMode hook
